Separation of test functions and code clean up
