+++ test/constants/constants.js
@@ -1,7 +1,25 @@
-const ERRORS = {
+module.exports = {
+    DEFAULT_PARAMETERS: {
+        settingsKerpak: {
+            USER_NAME_API: '14531737_api',
+            PASSWORD_API: 'ker1Vanadzor1pak1',
+            USER_NAME_API_BINDING: '14531737_binding',
+            PASSWORD_API_BINDING: 'ker1Vanadzor1pak1',
+            TEST_MODE: false
+        },
+        settings: {
+            USER_NAME_API: '14531661_api',
+            PASSWORD_API: 'ker1Vanadzor1pak',
+            USER_NAME_API_BINDING: '14531661_binding',
+            PASSWORD_API_BINDING: 'ker1Vanadzor1pak',
+            TEST_MODE: false
+        },
+    },
     TEST_NAMES: {
         ATTACH_CARD: 'Attach card',
+        SETTINGS: 'Test settings fields',
         SETTINGS_FIELDS: 'Testing without settings fields',
+        SETTINGS_WRONG: 'Testing incorrectly entered settings',
         TIMEOUT: 'Testing timeout',
         AMOUNT: 'Testing without amount',
         ORDER_NUMBER: 'Testing without order number',
@@ -40,6 +58,46 @@ const ERRORS = {
             'PASSWORD_API_BINDING is mandatory'
         ],
     },
+    ERROR_STEPS: {
+        FREEZE_STEP: 'registerPreAuth.do',
+        DEPOSIT_STEP: 'deposit.do',
+        REVERSE_STEP: 'reverse.do',
+        REGISTER_STEP: 'register.do',
+        REFUND_STEP: 'refund.do',
+        GET_BINDING_STEP: 'getBindings.do'
+    },
+    ACCESS_DENIED_REGISTER: {
+        hasError: true,
+        data: {
+            errorCode: 5,
+            errorCodeString: '5',
+            error: true
+        },
+        errorStep: 'register.do'
+    },
+    ACCESS_DENIED: {
+        hasError: true,
+        data: {
+            errorCode: '5',
+            error: true
+        }
+    },
+    ORDER_NO_FOUND: {
+        hasError: true,
+        data: {
+            errorCode: '6',
+            errorMessage: 'No such order',
+            error: true
+        }
+    },
+    CARD_SUCCESS: {
+        hasError: false,
+        data: {
+            errorCode: '0',
+            errorMessage: 'Success',
+            error: false
+        }
+    },
     TIMEOUT: {
         MESSAGE: 'TimeoutError',
     },
@@ -50,8 +108,7 @@ const ERRORS = {
             errorMessage: 'Amount is invalid',
             errorCodeString: '5',
             error: true
-        },
-        errorStep: 'register.do'
+        }
     },
     ORDER_NUMBER: {
         hasError: true,
@@ -60,8 +117,7 @@ const ERRORS = {
             errorMessage: 'Order number is duplicated, order with given order number is processed already',
             errorCodeString: '1',
             error: true
-        },
-        errorStep: 'register.do'
+        }
     },
     LANGUAGE_INVALID: {
         hasError: true,
@@ -86,7 +142,7 @@ const ERRORS = {
             error: false,
         }
     },
-    PAY_ORDER_NO_BINDING_FOUND: {
+    NO_BINDING_FOUND: {
         hasError: true,
         data: { error: 'No binding found', errorCode: 2 },
         errorStep: 'paymentOrderBinding.do'
@@ -121,26 +177,6 @@ const ERRORS = {
             error: false,
         }
     },
-    FREEZE_AMOUNT_INVALID: {
-        hasError: true,
-        data: {
-          errorCode: 5,
-          errorMessage: 'Amount is invalid',
-          errorCodeString: '5',
-          error: true
-        },
-        errorStep: 'registerPreAuth.do'
-    },
-    FREEZE_ORDER_NUMBER: {
-        hasError: true,
-        data: {
-          errorCode: 1,
-          errorMessage: 'Order number is duplicated, order with given order number is processed already',
-          errorCodeString: '1',
-          error: true
-        },
-        errorStep: 'registerPreAuth.do'
-    },
     FREEZE_LANGUAGE: {
         hasError: true,
         data: {
@@ -151,11 +187,6 @@ const ERRORS = {
         },
         errorStep: 'registerPreAuth.do'
     },
-    FREEZE_NO_BINDING_FOUND: {
-        hasError: true,
-        data: { error: 'No binding found', errorCode: 2 },
-        errorStep: 'paymentOrderBinding.do'
-    },
     REVERSE_SUCCESS: {
         reverse: { errorCode: '0', errorMessage: 'Success', error: false },
         hasError: false,
@@ -165,11 +196,6 @@ const ERRORS = {
           orderStatus: 3,
           error: false
         }
-    },
-    REVERSE_ORDER_NO_FOUND: {
-        hasError: true,
-        data: { errorCode: '6', errorMessage: 'No such order', error: true },
-        errorStep: 'reverse.do'
     },
     REVERSE_SAME_ORDER_ID: {
         hasError: true,
@@ -189,11 +215,6 @@ const ERRORS = {
           orderStatus: 2,
           error: false
         }
-    },
-    DEPOSIT_ORDER_NO_FOUND: {
-        hasError: true,
-        data: { errorCode: '6', errorMessage: 'No such order', error: true },
-        errorStep: 'deposit.do'
     },
     DEPOSIT_AMOUNT_INVALID: {
         hasError: true,
@@ -221,15 +242,6 @@ const ERRORS = {
           error: true
         },
         errorStep: 'deposit.do'
-    },
-    REFUND_ORDER_NO_FOUND: {
-        hasError: true,
-        data: {
-            errorCode: '6',
-            errorMessage: 'No such order',
-            error: true
-        },
-        errorStep: 'refund.do'
     },
     REFUND_SUCCESS: {
         refund: { errorCode: '0', errorMessage: 'Success', error: false },
@@ -268,14 +280,6 @@ const ERRORS = {
         },
         errorStep: 'refund.do'
     },
-    BINDING_SUCCESS: {
-        hasError: false,
-        data: {
-          errorCode: '0',
-          errorMessage: 'Success',
-          error: false
-        }
-    },
     BINDING_NOT_FOUND: {
         hasError: true,
         data: {
@@ -285,14 +289,6 @@ const ERRORS = {
           error: true
         },
         errorStep: 'getBindings.do'
-    },
-    REMOVE_CARD_SUCCESS: {
-        hasError: false,
-        data: {
-            errorCode: '0',
-            errorMessage: 'Success',
-            error: false
-        }
     },
     REMOVE_CARD_BINDING_NOT_FOUND: {
         hasError: true,
@@ -311,7 +307,14 @@ const ERRORS = {
           error: true
         },
         errorStep: 'unBindCard.do'
+    },
+    REMOVE_CARD_ACCESS_DENIED: {
+        hasError: true,
+        data: {
+            errorCode: '2',
+            errorMessage: 'No rights to manage binding of another merchant',
+            error: true
+        },
+        errorStep: 'unBindCard.do'
     }
 }
-
-module.exports = ERRORS
