+++ test/testFunctions/index.js
@@ -1,0 +1,11 @@
+module.exports = {
+    attachCardTest: require('./attachCard'),
+    payOrderTest: require('./payOrder'),
+    getOrderStatusTest: require('./getOrderStatus'),
+    freezeTest: require('./freeze'),
+    reverseOrderTest: require('./reverse'),
+    depositOrderTest: require('./deposit'),
+    refundOrderTest: require('./refund'),
+    getBindingsTest: require('./getBindings'),
+    removeCardTest: require('./removeCard')
+}
